refactor(search): extract TemplateSearchResultItem component
